Protect the deletes

+++ MyWebsite/MyWebsite/Controllers/ProjectsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using MyWebsite.Data;
 using MyWebsite.Models;
+using Microsoft.IdentityModel.Tokens;
 
 namespace MyWebsite.Controllers
 {
@@ -86,6 +87,7 @@ namespace MyWebsite.Controllers
 
             // Saves the data of the object in the session
             HttpContext.Session.SetInt32("ProjectId", projects.Id);
+            HttpContext.Session.SetString("ProjectTitle", projects.Title);
         }
 
         // POST: Projects/Edit/5
@@ -104,8 +106,9 @@ namespace MyWebsite.Controllers
 
             // Are the data I received from the session corresponding to the object sent to the browser?
             var ProjectId = HttpContext.Session.GetInt32("ProjectId");
+            var ProjectTitle = HttpContext.Session.GetString("ProjectTitle");
 
-            if (ProjectId == null)
+            if (ProjectId == null || ProjectTitle.IsNullOrEmpty())
             {
                 // Took too long to get the data from the session, or the session has expired
                 ModelState.AddModelError(string.Empty, "Project ID not found in session." + "You must restart the process");
@@ -114,7 +117,7 @@ namespace MyWebsite.Controllers
             }
 
             // Was there any adultation of the object?
-            if (projects.Id != ProjectId)
+            if (projects.Id != ProjectId || ProjectTitle != "Projects/Edit")
             {
                 // The user is trying to edit a different project than the one that was loaded in the session
                 return RedirectToAction("Index");
@@ -159,6 +162,10 @@ namespace MyWebsite.Controllers
                 return NotFound();
             }
 
+            // Saves the data of the object in the session
+            HttpContext.Session.SetInt32("ProjectId", projects.Id);
+            HttpContext.Session.SetString("ProjectTitle", projects.Title);
+
             return View(projects);
         }
 
@@ -168,11 +175,27 @@ namespace MyWebsite.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var projects = await _context.Projects.FindAsync(id);
+            var ProjectId = HttpContext.Session.GetInt32("ProjectId");
+            var ProjectTitle = HttpContext.Session.GetString("ProjectTitle");
+
+            // Are the data I received from the session corresponding to the object sent to the browser?
+            if (ProjectId == null || ProjectTitle.IsNullOrEmpty())
+            {
+                // Took too long to get the data from the session, or the session has expired
+                ModelState.AddModelError(string.Empty, "Project ID not found in session." + "You must restart the process");
+                return View(projects);
+            }
+
+            if (projects.Id != ProjectId || ProjectTitle != "Projects/Delete")
+            {
+                // The user is trying to delete a different project than the one that was loaded in the session
+                return RedirectToAction("Index");
+            }
+
             if (projects != null)
             {
                 _context.Projects.Remove(projects);
             }
-
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
